json as separate codec
